Expose synthesis results from synth builder (fix #172)
