+++ lib/src/synthesizers/synth_builder.dart
@@ -8,6 +8,7 @@
 /// Author: Max Korbel <max.korbel@intel.com>
 ///
 
+import 'package:collection/collection.dart';
 import 'package:rohd/rohd.dart';
 import 'package:rohd/src/utilities/uniquifier.dart';
 
@@ -26,6 +27,10 @@ class SynthBuilder {
 
   /// All the [SynthesisResult]s generated by this [SynthBuilder].
   final Set<SynthesisResult> _synthesisResults = {};
+
+  /// All the [SynthesisResult]s generated by this [SynthBuilder].
+  Set<SynthesisResult> get synthesisResults =>
+      UnmodifiableSetView(_synthesisResults);
 
   /// [Uniquifier] for instance type names.
   final Uniquifier _instanceTypeUniquifier = Uniquifier();
@@ -53,13 +58,9 @@ class SynthBuilder {
 
   /// Collects a [List] of [String]s representing file contents generated by
   /// the [synthesizer].
-  List<String> getFileContents() {
-    final fileContents = <String>{};
-    for (final synthesisResult in _synthesisResults) {
-      fileContents.add(synthesisResult.toFileContents());
-    }
-    return fileContents.toList();
-  }
+  List<String> getFileContents() => synthesisResults
+      .map((synthesisResult) => synthesisResult.toFileContents())
+      .toList();
 
   /// Provides an instance type name for [module].
   ///
