+++ lib/src/synthesizers/synthesizers.dart
@@ -1,5 +1,6 @@
-/// Copyright (C) 2021 Intel Corporation
+/// Copyright (C) 2021-2022 Intel Corporation
 /// SPDX-License-Identifier: BSD-3-Clause
 export 'synth_builder.dart';
 export 'synthesizer.dart';
 export 'systemverilog.dart';
+export 'synthesis_result.dart';
